refactor: rename component Rlates #79'
